+++ src/pytest_workflow/plugin.py
@@ -16,6 +16,7 @@
 
 """core functionality of pytest-workflow plugin"""
 import argparse
+import os
 import shutil
 import tempfile
 import warnings
@@ -72,7 +73,13 @@ def pytest_addoption(parser: PytestParser):
              "This ignores the .git directory, any untracked files and any "
              "files listed by .gitignore. "
              "Highly recommended when working in a git project.")
-
+    parser.addoption(
+        "--sb", "--stderr-bytes",
+        dest="stderr_bytes",
+        default=1000,
+        type=int,
+        help="The number of bytes to display from the stderr and "
+             "stdout on exitcode.")
     # Why `--tag <tag>` and not simply use `pytest -m <tag>`?
     # `-m` uses a "mark expression". So you have to type a piece of python
     # code instead of just supplying the tags you want. This is fine for the
@@ -448,7 +455,8 @@ class WorkflowTestsCollector(pytest.Collector):
         tests += [ExitCodeTest.from_parent(
             parent=self,
             desired_exit_code=self.workflow_test.exit_code,
-            workflow=workflow)]
+            workflow=workflow,
+            stderr_bytes=self.config.getoption("stderr_bytes"))]
 
         tests += [ContentTestCollector.from_parent(
             name="stdout", parent=self,
@@ -470,9 +478,10 @@ class WorkflowTestsCollector(pytest.Collector):
 class ExitCodeTest(pytest.Item):
     def __init__(self, parent: pytest.Collector,
                  desired_exit_code: int,
-                 workflow: Workflow):
+                 workflow: Workflow, stderr_bytes: int):
         name = f"exit code should be {desired_exit_code}"
         super().__init__(name, parent=parent)
+        self.stderr_bytes = stderr_bytes
         self.workflow = workflow
         self.desired_exit_code = desired_exit_code
 
@@ -481,7 +490,18 @@ class ExitCodeTest(pytest.Item):
         assert self.workflow.exit_code == self.desired_exit_code
 
     def repr_failure(self, excinfo, style=None):
-        message = (f"'{self.workflow.name}' exited with exit code " +
-                   f"'{self.workflow.exit_code}' instead of "
-                   f"'{self.desired_exit_code}'.")
+        standerr = self.workflow.stderr_file
+        standout = self.workflow.stdout_file
+        with open(standout, "rb") as standout_file, \
+             open(standerr, "rb") as standerr_file:
+            if os.path.getsize(standerr) >= self.stderr_bytes:
+                standerr_file.seek(-self.stderr_bytes, os.SEEK_END)
+            if os.path.getsize(standout) >= self.stderr_bytes:
+                standout_file.seek(-self.stderr_bytes, os.SEEK_END)
+            message = (f"'{self.workflow.name}' exited with exit code " +
+                       f"'{self.workflow.exit_code}' instead of "
+                       f"'{self.desired_exit_code}'.\nstderr: "
+                       f"{standerr_file.read().strip().decode('utf-8')}"
+                       f"\nstdout: "
+                       f"{standout_file.read().strip().decode('utf-8')}")
         return message