+++ tests/test_fail_messages.py
@@ -29,6 +29,11 @@ FAILURE_MESSAGE_TESTS: List[Tuple[str, str]] = [
     """,
      "'fail_test' exited with exit code '2' instead of '0'"),
     ("""\
+    - name: exitcode_test
+      command: bash -c 'printf "This code had an error" >&2 ; exit 12'
+    """,
+     "stderr: This code had an error"),
+    ("""\
     - name: create file
       command: echo moo
       files:
@@ -169,4 +174,27 @@ def test_messages(test: str, message: str, pytester):
     # Ideally this should be run in a LC_ALL=C environment. But this is not
     # possible due to multiple levels of process launching.
     result = pytester.runpytest("-v")
+    assert message in result.stdout.str()
+
+
+EXITCODE_MESSAGE_TESTS: List[Tuple[str, str]] = [
+    ("""\
+    - name: stderr_bytes_stderr_test
+      command: bash -c 'printf "This code had an error" ; exit 12'
+    """,
+     "stdout: error"),
+    ("""\
+    - name: stderr_bytes_stdout_test
+      command: bash -c 'printf "This code had an error" >&2 ; exit 12'
+    """,
+     "stderr: error")
+]
+
+
+@pytest.mark.parametrize(["test", "message"], EXITCODE_MESSAGE_TESTS)
+def test_messages_exitcode(test: str, message: str, pytester):
+    pytester.makefile(".yml", textwrap.dedent(test))
+    # Ideally this should be run in a LC_ALL=C environment. But this is not
+    # possible due to multiple levels of process launching.
+    result = pytester.runpytest("-v", "--sb", "5")
     assert message in result.stdout.str()